monkeychenTSL/Python_Tsl/TSL_work/requirements.txt
allure-pytest
jsonpath
pyyaml
pytest-yaml
requests
pytest-xlsx
pytest
pytest-yaml-sanmu
selenium
